+++ client/internal/routemanager/systemops_bsd_nonios.go
@@ -1,0 +1,13 @@
+//go:build (darwin || dragonfly || freebsd || netbsd || openbsd) && !ios
+
+package routemanager
+
+import "net/netip"
+
+func addToRouteTableIfNoExists(prefix netip.Prefix, addr string, intf string) error {
+	return genericAddToRouteTableIfNoExists(prefix, addr, intf)
+}
+
+func removeFromRouteTableIfNonSystem(prefix netip.Prefix, addr string, intf string) error {
+	return genericRemoveFromRouteTableIfNonSystem(prefix, addr, intf)
+}
