+++ client/internal/routemanager/systemops_nonlinux_test.go
@@ -1,5 +1,4 @@
-//go:build !linux
-// +build !linux
+//go:build !linux || android
 
 package routemanager
