+++ util/net/listener_generic.go
@@ -1,4 +1,4 @@
-//go:build !linux
+//go:build !linux || android
 
 package net
